Split into httpann.py and httpann_example.py. Made names more consistent.
