+++ algos.yaml
@@ -45,9 +45,9 @@ random-range-xs:
               [{"R":32, "L":32, "B":0.0001, "M":1}]
           query-args: |
               [{"Ls":10, "BW":4, "T":16}]
-    http-ann-example:
-      docker-tag: billion-scale-benchmark-httpannexample
-      module: benchmark.algorithms.httpann
+    httpann_example:
+      docker-tag: billion-scale-benchmark-httpann_example
+      module: benchmark.algorithms.httpann_example
       constructor: HttpANNExampleAlgorithm
       base-args: [ "@metric", "@dimension" ]
       run-groups:
@@ -83,9 +83,9 @@ random-xs:
               [{"R":32, "L":32, "B":0.0001, "M":1}]
           query-args: |
               [{"Ls":10, "BW":4, "T":16}]
-    http-ann-example:
-      docker-tag: billion-scale-benchmark-httpannexample
-      module: benchmark.algorithms.httpann
+    httpann_example:
+      docker-tag: billion-scale-benchmark-httpann_example
+      module: benchmark.algorithms.httpann_example
       constructor: HttpANNExampleAlgorithm
       base-args: [ "@metric", "@dimension" ]
       run-groups:
