+++ benchmark/algorithms/httpann.py
@@ -6,11 +6,8 @@ from threading import Thread
 
 import numpy as np
 import requests
-from flask import jsonify
-from sklearn.neighbors import NearestNeighbors
 
 from benchmark.algorithms.base import BaseANN
-from benchmark.datasets import DATASETS
 
 
 class HttpANN(BaseANN):
@@ -35,6 +32,13 @@ class HttpANN(BaseANN):
     """
 
     def __init__(self, server_url: str, start_seconds: int, name: str, **kwargs):
+        """
+        Base constructor for an HttpANN algorithm.
+        @param server_url: base URL for the server including port, e.g., "http:localhost:8080"
+        @param start_seconds: how many seconds to wait for the server to start before posting to the /init endpoint.
+        @param name: algorithm name
+        @param kwargs: any additional keyword arguments that will be passed through to the /init endpoint.
+        """
         self.server_url = server_url
         self.name = name
 
@@ -94,6 +98,7 @@ class HttpANNError(RuntimeError):
 
 class HttpANNResponseError(HttpANNError):
     """Custom error type"""
+
     def __init__(self, endpoint: str, expected_status: int, actual_status: int):
         super(HttpANNError, self).__init__(f"Endpoint {endpoint} expected {expected_status} but got {actual_status}")
 
@@ -104,6 +109,7 @@ class HttpANNSubprocess(object):
     Starts a background thread to monitor the subprocess by checking for an exit code once per second.
     If the background thread finds an exit code, it will raise an HttpANNError.
     """
+
     def __init__(self, server_subprocess_command: str):
         proc = Popen(shlex.split(server_subprocess_command), stdout=sys.stdout, stderr=sys.stderr)
 
@@ -116,150 +122,3 @@ class HttpANNSubprocess(object):
 
         t = Thread(target=monitor, args=(), daemon=True)
         t.start()
-
-
-class HttpANNExampleAlgorithm(HttpANN, HttpANNSubprocess):
-    """
-    ANN algorithm that serves as a standard "algorithm" (callable from runner.py) and manages an HTTP server that
-    implements the actual indexing and query processing algorithms.
-
-    By implementing HttpANNSubprocess, it starts a local server (which is implemented further below in the same file).
-    By implementing HttpANN, it can be used by runner.py to make ANN requests from Python to the local server.
-
-    Obviously this is a contrived setup, as the actual algorithm is also implemented in Python.
-    It's purely as an example of how one might run an algorithm from another language by using an HTTP server to
-    implement the server API expected by HttpANN.
-    """
-    def __init__(self, metric: str, dimension: int, use_dims: float):
-        HttpANNSubprocess.__init__(self, "python3 -m benchmark.algorithms.httpann example")
-        HttpANN.__init__(self, server_url="http://localhost:8080", start_seconds=3,
-                         name=f"http-ann-example-{metric}-{use_dims}", metric=metric, dimension=dimension,
-                         use_dims=use_dims)
-
-
-# Starts a local flask server that adheres to the HttpANN API and delegates the work to a local ANN algorithm.
-if __name__ == "__main__" and sys.argv[-1] == "example":
-
-    class SimpleANNAlgo(object):
-        """
-        Very simple ANN algorithm intended only to demonstrate the HttpANN functionality.
-        This algorithm is instantiated and called from the example server below.
-        The algorithm is approximate in the sense that it uses exact KNN constrained to a configurable subset of the
-        highest variance dimensions. For example, if dimensions=100 and use_dims=0.22, the algorithm picks the 22
-        dimensions with the highest variance and use them for exact KNN.
-        """
-
-        def __init__(self, metric: str, dimension: int, use_dims: float = 0.1):
-            self.metric = metric
-            self.dimension = dimension
-            self.use_dims = use_dims
-            self.knn = None
-            self.high_variance_dims = None
-            self.res = None
-
-        def fit(self, dataset):
-            ds = DATASETS[dataset]()
-            arr = ds.get_dataset()
-            var = arr.var(axis=0)
-            num_dims = int(self.use_dims * arr.shape[1])
-            self.high_variance_dims = np.argsort(var)[-num_dims:]
-            self.knn = NearestNeighbors(algorithm='brute', metric=self.metric)
-            self.knn.fit(arr[:, self.high_variance_dims])
-
-        def load_index(self, dataset):
-            # Always returns false because the index is not stored.
-            return False
-
-        def query(self, X, k):
-            self.res = self.knn.kneighbors(X[:, self.high_variance_dims], n_neighbors=k, return_distance=False)
-
-        def range_query(self, X, radius):
-            nbrs, dsts = self.knn.radius_neighbors(X[:, self.high_variance_dims], radius=radius, return_distance=True)
-            total = sum(map(len, nbrs))
-            lims = np.zeros(len(X) + 1, 'int32')
-            I = np.zeros(total, 'int32')
-            D = np.zeros(total, 'float32')
-            for i in range(len(X)):
-                lims[i + 1] = lims[i] + len(nbrs[i])
-                I[lims[i]:lims[i + 1]] = nbrs[i]
-                D[lims[i]:lims[i + 1]] = dsts[i]
-            self.res = (lims, I, D)
-
-        def get_results(self):
-            return self.res
-
-        def get_range_results(self):
-            return self.res
-
-        def get_additional(self):
-            return {}
-
-
-    from flask import Flask, request
-
-    app = Flask(__name__)
-
-    # Algorithm is instantiated later but needs to be attached to an object.
-    app.algo = None
-
-    @app.route("/status", methods=['GET'])
-    def status():
-        return jsonify(dict()), 200
-
-    @app.route("/init", methods=['POST'])
-    def init():
-        app.algo = SimpleANNAlgo(**request.json)
-        return jsonify(dict()), 200
-
-    @app.route("/load_index", methods=['POST'])
-    def load_index():
-        b = app.algo.load_index(**request.json)
-        return jsonify(dict(load_index=b)), 200
-
-    @app.route("/fit", methods=['POST'])
-    def fit():
-        app.algo.fit(**request.json)
-        return jsonify(dict()), 200
-
-    @app.route("/set_query_arguments", methods=['POST'])
-    def set_query_arguments():
-        app.algo.set_query_arguments(**request.json)
-        return jsonify(dict()), 200
-
-    @app.route("/query", methods=['POST'])
-    def query():
-        j = request.json
-        app.algo.query(np.array(j['X']), j['k'])
-        return jsonify(dict()), 200
-
-    @app.route("/range_query", methods=['POST'])
-    def range_query():
-        j = request.json
-        app.algo.range_query(np.array(j['X']), j['radius'])
-        return jsonify(dict()), 200
-
-    @app.route("/get_results", methods=['POST'])
-    def get_results():
-        neighbors = [arr.tolist() for arr in app.algo.res]
-        return jsonify(dict(get_results=neighbors)), 200
-
-    @app.route("/get_range_results", methods=['POST'])
-    def get_range_results():
-        lims, I, D = app.algo.get_range_results()
-        res = [
-            lims.tolist(),
-            [arr.tolist() for arr in I],
-            [arr.tolist() for arr in D]
-        ]
-        return jsonify(dict(get_range_results=res)), 200
-
-    @app.route("/get_additional", methods=['POST'])
-    def get_additional():
-        return jsonify(dict(get_additional=app.algo.get_additional())), 200
-
-    app.run('0.0.0.0', 8080, debug=False)
-    # We could also use gevent/wsgi for a more professional setup.
-    # https://flask.palletsprojects.com/en/2.0.x/deploying/wsgi-standalone/#gevent
-    # from gevent.pywsgi import WSGIServer
-    # http_server = WSGIServer(('0.0.0.0', 8080), app)
-    # http_server.serve_forever()
